requirements.txt
# Existing Repository Requirements
langchain~=0.3.25
openai==0.28.1
pydantic>=1.10.12
pytest>=7.4.0
transformers>=4.31.0
chromadb==0.4.12
tiktoken==0.4.0
pandas~=2.2.3
python-dotenv~=1.1.0